feat(cli): move CLI implementations into importable chiq.cli package
