+++ python/package/chiq/cli/__init__.py
@@ -1,0 +1,1 @@
+# chiq CLI entry-point modules
